fix incorrect js template;  make newsignin, reervation, resetpassword templates narrower and add border
