+++ packages/cli/src/generator_templates/js/emails/NewSignIn.jsx
@@ -16,16 +16,23 @@ import {
   MjmlColumn,
   MjmlText,
   MjmlSpacer,
+  MjmlDivider,
 } from "mjml-react";
 
 const NewSignIn = ({ name, headline, body, bulletedList }) => {
   return (
     <Mjml>
       <Head />
-      <MjmlBody width={600}>
+      <MjmlBody width={400}>
         <Header />
         <MjmlSection padding="0 24px" cssClass="smooth">
           <MjmlColumn>
+          <MjmlDivider
+              borderColor="#666"
+              borderStyle="dotted"
+              borderWidth="1px"
+              padding="8px 0"
+            ></MjmlDivider>
             <MjmlText
               padding="24px 0 8px"
               fontSize={textLg}
@@ -64,6 +71,12 @@ const NewSignIn = ({ name, headline, body, bulletedList }) => {
               <br />
               The BookBook Team
             </MjmlText>
+            <MjmlDivider
+              borderColor="#666"
+              borderStyle="dotted"
+              borderWidth="1px"
+              padding="20px 0 8px"
+            ></MjmlDivider>
           </MjmlColumn>
         </MjmlSection>
         <Footer />
