+++ packages/cli/src/generator_templates/js/emails/Reservation.jsx
@@ -16,18 +16,25 @@ import {
   MjmlColumn,
   MjmlText,
   MjmlSpacer,
+  MjmlDivider,
 } from "mjml-react";
 
 const Reservation = ({ headline, body, bulletedList, ctaText }) => {
   return (
     <Mjml>
       <Head />
-      <MjmlBody width={600}>
+      <MjmlBody width={400}>
         <Header />
         <MjmlSection padding="0 24px" cssClass="smooth">
           <MjmlColumn>
+          <MjmlDivider
+              borderColor="#666"
+              borderStyle="dotted"
+              borderWidth="1px"
+              padding="8px 0"
+            ></MjmlDivider>
             <MjmlText
-              padding="24px 0 8px"
+              padding="24px 0"
               fontSize={textLg}
               lineHeight={leadingTight}
               cssClass="paragraph"
@@ -36,13 +43,6 @@ const Reservation = ({ headline, body, bulletedList, ctaText }) => {
             </MjmlText>
             {bulletedList}
             <MjmlSpacer height="16px" />
-            {ctaText && (
-              <>
-                <MjmlSpacer height="8px" />
-                <ButtonPrimary link={"#"} uiText={ctaText} />
-                <MjmlSpacer height="24px" />
-              </>
-            )}
             <MjmlText
               cssClass="paragraph"
               padding="0"
@@ -53,16 +53,19 @@ const Reservation = ({ headline, body, bulletedList, ctaText }) => {
               {body}
               </>
             </MjmlText>
-            <MjmlText
-              padding="16px 0"
-              fontSize={textBase}
-              lineHeight={leadingRelaxed}
-              cssClass="paragraph"
-            >
-              ♥,
-              <br />
-              The BookBook Team
-            </MjmlText>
+            {ctaText && (
+              <>
+                <MjmlSpacer height="24px" />
+                <ButtonPrimary link={"#"} uiText={ctaText} />
+                <MjmlSpacer height="8px" />
+              </>
+            )}
+            <MjmlDivider
+              borderColor="#666"
+              borderStyle="dotted"
+              borderWidth="1px"
+              padding="32px 0 0"
+            ></MjmlDivider>
           </MjmlColumn>
         </MjmlSection>
         <Footer />
