+++ packages/cli/src/generator_templates/ts/emails/components/ButtonPrimary.tsx
@@ -13,7 +13,7 @@ const ButtonPrimary: React.FC<ButtonPrimaryProps> = ({ link, uiText }) => {
       <MjmlButton
         lineHeight={leadingTight}
         fontSize={textBase}
-        height={52}
+        height={32}
         padding="0"
         align="left"
         href={link}
@@ -27,7 +27,7 @@ const ButtonPrimary: React.FC<ButtonPrimaryProps> = ({ link, uiText }) => {
       <MjmlButton
         lineHeight={leadingTight}
         fontSize={textBase}
-        height={52}
+        height={32}
         padding="0"
         align="left"
         href={link}
